refactor: reorganize files and directories
