--- src/components/Welcome.astro
@@ -1,5 +1,0 @@
----
-// import Layout from '@/layouts/Layout.astro';
----
-
-<h1>Welcome</h1>
