+++ src/modules/core/layouts/Layout.astro
@@ -1,6 +1,6 @@
 ---
-import Header from '@/components/header.astro';
 import '@/styles/global.css';
+import Header from '@/modules/core/layouts/header.astro';
 ---
 <!doctype html>
 <html lang="en">
