+++ src/pages/auth/login.astro
@@ -1,9 +1,9 @@
 ---
-import Layout from '@/layouts/Layout.astro';
-import { Label } from '@/components/ui/label';
-import { Input } from '@/components/ui/input';
-import { Button } from '@/components/ui/button';
-import { Link } from '@/components/ui/link';
+import Layout from '@/modules/core/layouts/Layout.astro';
+import { Label } from '@/modules/core/ui/label';
+import { Input } from '@/modules/core/ui/input';
+import { Button } from '@/modules/core/ui/button';
+import { Link } from '@/modules/core/ui/link';
 ---
 <Layout>
   <main class="flex flex-col gap-12 items-center mt-10">
