+++ src/pages/index.astro
@@ -1,10 +1,6 @@
 ---
-import MyDashboard from '@/components/page/studentProgress/MyDashboard.astro';
-// import Welcome from '../components/Welcome.astro';
-import Layout from '../layouts/Layout.astro';
-
-// Welcome to Astro! Wondering what to do next? Check out the Astro documentation at https://docs.astro.build
-// Don't want to use any of this? Delete everything in this file, the `assets`, `components`, and `layouts` directories, and start fresh.
+import Layout from '@/modules/core/layouts/Layout.astro';
+import MyDashboard from '@/modules/courses/components/MyDashboard.astro';
 ---
 
 <Layout>
